python/bots/requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.0
async-timeout==3.0.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.3
Click==7.0
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
dbus-python==1.2.16
discord.py==2.2.2
distro==1.4.0
entrypoints==0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.9.0
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
iso-639==0.4.5
iso3166==1.0.1
isodate==0.6.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.2
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
matplotlib==3.4.3
more-itertools==4.2.0
multidict==5.1.0
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.1.0
openai==0.27.2
opencv-python
packaging==21.0
peewee==3.15.1
pexpect==4.6.0
Pillow==9.4.0
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.10.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pytest==6.2.4
pytest-asyncio==0.15.1
pytz==2021.1
PyYAML==5.3.1
requests==2.28.2
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
ssh-import-id==5.10
